Reorgmanize the folder structure, added cmd and internal folders, and moved files to conform to idiomatic folder structure in Go
