+++ internal/todo/utils.go
@@ -1,4 +1,4 @@
-package main
+package todo
 
 import (
 	"bufio"
